docs(modal): update Modal docs
